+++ client/src/pages/Home.jsx
@@ -10,6 +10,7 @@ import useFetch from '../util/useFetch';
 import PageButton from '../components/buttons/PageButton';
 import { useLocation } from 'react-router-dom';
 import BASE_URL from '../constants/baseUrl';
+import handleAuthError from '../exception/handleAuthError';
 
 function Home() {
   const { search } = useLocation();
@@ -31,28 +32,7 @@ function Home() {
       return;
     }
 
-    if (response.status === 401) {
-      alert('로그인 후 질문을 작성할 수 있습니다.');
-      window.location.href = ROUTE_PATH.LOGIN;
-
-      return;
-    }
-
-    if (response.status === 403) {
-      const response = await fetch(BASE_URL + '/auth/reissuetoken');
-
-      if (!response.ok) {
-        window.location.href = ROUTE_PATH.LOGIN;
-
-        return;
-      }
-
-      const authorization = response.headers.get('Authorization');
-      const refresh = response.headers.get('Refresh');
-
-      localStorage.setItem('Authorization', authorization);
-      localStorage.setItem('Refresh', refresh);
-    }
+    handleAuthError(response.status, handleConfirmLogin);
   }
 
   return (
